olicyberit-regional-ctf/web03/src/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
Werkzeug==3.0.1
gunicorn==21.2.0
pycryptodome==3.19.0